# core/byosnap-go/snapser-resources/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "description": "BYOSnap Core Go Example - minimal starter scaffold. Every endpoint Snapser expects is present, but each handler body is a stub. See advanced/byosnap-go for full implementations.",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title": "BYOSnap Core Go Example",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:5003/"
    }
  ],
  "paths": {
    "/v1/byosnap-core/example/admin": {
      "get": {
        "description": "Exposed over api-key + internal auth and surfaced in the Admin SDK via x-snapser-sdk-categories.",
        "operationId": "ExampleAdminSdk",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessageSchema"
                }
              }
            },
            "description": "Success"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            },
            "description": "Unauthorized access"
          }
        },
        "summary": "Example: Admin SDK",
        "x-snapser-auth-types": [
          "api-key",
          "internal"
        ],
        "x-snapser-sdk-categories": ["admin"]
      }
    },
    "/v1/byosnap-core/example/api-key": {
      "get": {
        "description": "Accessible with a valid API key. Use for trusted server-to-server calls.",
        "operationId": "ExampleApiKeyAuth",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessageSchema"
                }
              }
            },
            "description": "Success"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            },
            "description": "Unauthorized access"
          }
        },
        "summary": "Example: Api-Key Auth",
        "x-snapser-auth-types": [
          "api-key"
        ]
      }
    },
    "/v1/byosnap-core/example/internal": {
      "get": {
        "description": "Callable only by other Snaps within the same Snapend (internal gateway). Surfaces in the internal SDK.",
        "operationId": "ExampleInternalAuth",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessageSchema"
                }
              }
            },
            "description": "Success"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            },
            "description": "Unauthorized access"
          }
        },
        "summary": "Example: Internal Auth",
        "x-snapser-auth-types": [
          "internal"
        ]
      }
    },
...
    "/v1/byosnap-core/users/{user_id}/example": {
      "get": {
        "description": "Accessible by a logged-in user, validated against the User-Id in their token. Surfaces in the client/game SDK.",
        "operationId": "ExampleUserAuth",
        "parameters": [
          {
            "description": "Unique identifier of the user",
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessageSchema"
                }
              }
            },
            "description": "Success"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            },
            "description": "Unauthorized access"
          }
        },
        "summary": "Example: User Auth",
        "x-snapser-auth-types": [
          "user"
        ]
      }
    },
    "/v1/byosnap-core/users/{user_id}/example/multi-auth": {
      "get": {
        "description": "One endpoint reachable by a logged-in user, a valid API key, or an internal Snap. List every auth type you want to allow - no need for a separate route per type.",
        "operationId": "ExampleMultiAuth",
        "parameters": [
          {
            "description": "Unique identifier of the user",
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessageSchema"
                }
              }
            },
            "description": "Success"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseSchema"
                }
              }
            },
            "description": "Unauthorized access"
          }
        },
        "summary": "Example: Multi Auth",
        "x-snapser-auth-types": [
          "user",
          "api-key",
          "internal"
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponseSchema": {
        "properties": {
          "error_message": {
            "type": "string",
            "x-go-name": "ErrorMessage"
          }
        },
        "type": "object",
        "x-go-package": "byosnap-go"
      },
...
      "SuccessMessageSchema": {
        "properties": {
          "message": {
            "type": "string",
            "x-go-name": "Message"
          }
        },
        "type": "object",
        "x-go-package": "byosnap-go"
      },
...
    }
  },
...
}